requirements.txt
mysql-connector-python~=8.0.31
Flask~=2.2.2